New warning message styles and minor layout adjustments
- Added a new class `.outofdate-warning` in `assets/sass/posts.scss` to apply specific styles for warning messages.
- Updated the article element in `layouts/notes/list.html` to add `my-4` for vertical spacing between articles, enhancing readability and spacing consistency.

This commit introduces a new warning style to ensure that out-of-date content is clearly marked, helping users to quickly identify important information. Also, by adjusting the margins on the articles, we’ve improved the overall aesthetic and organization of note entries, making it less of a visual clusterfuck—because who needs that, am I right? 😎